Added gun script, Target script
Gun "Model" can be changed, as of right now it's just a simple rectangle/cube. We can change the model to something the Graphics students make later. The gun script handles the raycasting, the Target script has to be in anything we want to take damage (Players, Enemy AI, etc). The damage and range of the gun can be changed in the editor, and the health of the object we want to hit can be changed in the editor as well. (Matt wanted me to do the raycasting because he had other stuff to do).